frontend/requirements.txt
streamlit
requests
pillow
streamlit_drawable_canvas
psycopg2